+++ src/Lexer/Recognizer/RegexRecognizer.php
@@ -13,31 +13,19 @@ namespace Dissect\Lexer\Recognizer;
  */
 class RegexRecognizer implements Recognizer
 {
-    protected string $regex;
-
-    /**
-     * Constructor.
-     *
-     * @param string $regex The regex to use in the match.
-     */
-    public function __construct(string $regex)
-    {
-        $this->regex = $regex;
-    }
+    public function __construct(protected readonly string $regex) {}
 
     /**
      * {@inheritDoc}
      */
-    public function match(string $string, ?string &$result = null): bool
+    public function match(string $string): ?string
     {
         $r = preg_match($this->regex, $string, $match, PREG_OFFSET_CAPTURE);
 
         if ($r === 1 && $match[0][1] === 0) {
-            $result = $match[0][0];
-
-            return true;
+            return $match[0][0];
         }
 
-        return false;
+        return null;
     }
 }