+++ src/Lexer/RegexLexer.php
@@ -7,6 +7,7 @@ namespace Dissect\Lexer;
 use Dissect\Lexer\TokenStream\ArrayTokenStream;
 use Dissect\Lexer\TokenStream\TokenStream;
 use Dissect\Parser\Parser;
+use RuntimeException;
 
 /**
  * Fast regex lexer, adapted from Doctrine.
@@ -19,28 +20,33 @@ use Dissect\Parser\Parser;
  */
 abstract class RegexLexer implements Lexer
 {
+    private ?string $regex = null;
+
     /**
      * {@inheritDoc}
      */
     public function lex(string $string): TokenStream
     {
-        static $regex;
-
-        if (!isset($regex)) {
-            $regex = '/(' . implode(')|(', $this->getCatchablePatterns()) . ')|'
+        if ($this->regex === null) {
+            $this->regex = '/(' . implode(')|(', $this->getCatchablePatterns()) . ')|'
                 . implode('|', $this->getNonCatchablePatterns()) . '/i';
         }
 
         $string = strtr($string, ["\r\n" => "\n", "\r" => "\n"]);
 
         $flags = PREG_SPLIT_NO_EMPTY | PREG_SPLIT_DELIM_CAPTURE | PREG_SPLIT_OFFSET_CAPTURE;
-        $matches = preg_split($regex, $string, -1, $flags);
+        $matches = preg_split($this->regex, $string, -1, $flags);
+
+        if ($matches === false) {
+            throw new RuntimeException(sprintf('preg_split failed for regex: %s', $this->regex));
+        }
+
         $tokens = [];
         $line = 1;
         $oldPosition = 0;
 
         foreach ($matches as $match) {
-            list ($value, $position) = $match;
+            [$value, $position] = $match;
 
             $type = $this->getType($value);
 
@@ -60,11 +66,15 @@ abstract class RegexLexer implements Lexer
 
     /**
      * The patterns corresponding to tokens.
+     *
+     * @return string[]
      */
     abstract protected function getCatchablePatterns(): array;
 
     /**
      * The patterns corresponding to tokens to be skipped.
+     *
+     * @return string[]
      */
     abstract protected function getNonCatchablePatterns(): array;