+++ src/Node/Node.php
@@ -12,31 +12,25 @@ use RuntimeException;
  * A basic contract for a node in an AST.
  *
  * @author Jakub Lédl <jakubledl@gmail.com>
+ *
+ * @template-extends IteratorAggregate<int|string, Node>
  */
 interface Node extends Countable, IteratorAggregate
 {
     /**
      * Returns the children of this node.
      *
-     * @return array The children belonging to this node.
+     * @return array<int|string, Node>
      */
     public function getNodes(): array;
 
     /**
      * Checks for existence of child node named $name.
-     *
-     * @param string $name The name of the child node.
-     *
-     * @return boolean If the node exists.
      */
     public function hasNode(string $name): bool;
 
     /**
      * Returns a child node specified by $name.
-     *
-     * @param int|string $name The name of the node.
-     *
-     * @return Node The child node specified by $name.
      *
      * @throws RuntimeException When no child node named $name exists.
      */
@@ -44,40 +38,28 @@ interface Node extends Countable, IteratorAggregate
 
     /**
      * Sets a child node.
-     *
-     * @param string $name The name.
-     * @param Node $child The new child node.
      */
-    public function setNode(string $name, Node $child);
+    public function setNode(string $name, Node $child): void;
 
     /**
      * Removes a child node by name.
-     *
-     * @param string $name The name.
      */
-    public function removeNode(string $name);
+    public function removeNode(string $name): void;
 
     /**
      * Returns all attributes of this node.
      *
-     * @return array The attributes.
+     * @return array<string, mixed>
      */
     public function getAttributes(): array;
 
     /**
-     * Determines whether this node has an attribute
-     * under $key.
-     *
-     * @param string $key The key.
-     * @return boolean Whether there's an attribute under $key.
+     * Determines whether this node has an attribute under $key.
      */
     public function hasAttribute(string $key): bool;
 
     /**
      * Gets an attribute by key.
-     *
-     * @param string $key The key.
-     * @return mixed The attribute value.
      *
      * @throws RuntimeException When no attribute exists under $key.
      */
@@ -85,16 +67,11 @@ interface Node extends Countable, IteratorAggregate
 
     /**
      * Sets an attribute by key.
-     *
-     * @param string $key The key.
-     * @param mixed $value The new value.
      */
-    public function setAttribute(string $key, mixed $value);
+    public function setAttribute(string $key, mixed $value): void;
 
     /**
      * Removes an attribute by key.
-     *
-     * @param string $key The key.
      */
-    public function removeAttribute(string $key);
+    public function removeAttribute(string $key): void;
 }